Changes related to feature/UNIONVMS-2788
